delegator: move proxy build to utility func
